MAINT: Cleanup to break Matter dependencies
Where not required
