+++ client/HelperFunctions.cpp
@@ -11,9 +11,9 @@
 */
 #include "HelperFunctions.h"
 
-#include <iostream>
+#include "client/BaseStructures.h"
+#include <fstream>
 #include <math.h>
-#include <set>
 #include <string.h>
 #include <time.h>
 
@@ -244,129 +244,6 @@ RotationMatrix helper_functions::rotationExtract(const AtomMatrix r1,
   return R;
 }
 
-bool helper_functions::rotationMatch(const Matter &m1, const Matter &m2,
-                                     const double max_diff) {
-  AtomMatrix r1 = m1.getPositions();
-  AtomMatrix r2 = m2.getPositions();
-
-  // Align centroids
-  VectorType c1(3);
-  VectorType c2(3);
-
-  c1[0] = r1.col(0).sum();
-  c1[1] = r1.col(1).sum();
-  c1[2] = r1.col(2).sum();
-  c2[0] = r2.col(0).sum();
-  c2[1] = r2.col(1).sum();
-  c2[2] = r2.col(2).sum();
-  c1 /= r1.rows();
-  c2 /= r2.rows();
-
-  for (int i = 0; i < r1.rows(); i++) {
-    r1(i, 0) -= c1[0];
-    r1(i, 1) -= c1[1];
-    r1(i, 2) -= c1[2];
-
-    r2(i, 0) -= c2[0];
-    r2(i, 1) -= c2[1];
-    r2(i, 2) -= c2[2];
-  }
-
-  RotationMatrix R = rotationExtract(r1, r2);
-
-  // Eigen is transposed relative to numpy
-  r2 = r2 * R;
-
-  for (int i = 0; i < r1.rows(); i++) {
-    double diff = (r2.row(i) - r1.row(i)).norm();
-    if (diff > max_diff) {
-      return false;
-    }
-  }
-  return true;
-}
-
-void helper_functions::rotationRemove(const AtomMatrix r1_passed,
-                                      std::shared_ptr<Matter> m2) {
-  AtomMatrix r1 = r1_passed;
-  AtomMatrix r2 = m2->getPositions();
-
-  // Align centroids
-  VectorType c1(3);
-  VectorType c2(3);
-
-  c1[0] = r1.col(0).sum();
-  c1[1] = r1.col(1).sum();
-  c1[2] = r1.col(2).sum();
-  c2[0] = r2.col(0).sum();
-  c2[1] = r2.col(1).sum();
-  c2[2] = r2.col(2).sum();
-  c1 /= r1.rows();
-  c2 /= r2.rows();
-
-  for (int i = 0; i < r1.rows(); i++) {
-    r1(i, 0) -= c1[0];
-    r1(i, 1) -= c1[1];
-    r1(i, 2) -= c1[2];
-
-    r2(i, 0) -= c2[0];
-    r2(i, 1) -= c2[1];
-    r2(i, 2) -= c2[2];
-  }
-
-  RotationMatrix R = rotationExtract(r1, r2);
-
-  // Eigen is transposed relative to numpy
-  r2 = r2 * R;
-
-  // Move centroid back to initial position
-  for (int i = 0; i < r2.rows(); i++) {
-    r2(i, 0) += c2[0];
-    r2(i, 1) += c2[1];
-    r2(i, 2) += c2[2];
-  }
-
-  m2->setPositions(r2);
-  return;
-}
-
-void helper_functions::rotationRemove(const std::shared_ptr<Matter> m1,
-                                      std::shared_ptr<Matter> m2) {
-  AtomMatrix r1 = m1->getPositions();
-  rotationRemove(r1, m2);
-  return;
-}
-
-void helper_functions::translationRemove(Matter &m1,
-                                         const AtomMatrix r2_passed) {
-  AtomMatrix r1 = m1.getPositions();
-  AtomMatrix r2 = r2_passed;
-
-  // net displacement
-  VectorType disp(3);
-  AtomMatrix r12 = m1.pbc(r2 - r1);
-
-  disp[0] = r12.col(0).sum();
-  disp[1] = r12.col(1).sum();
-  disp[2] = r12.col(2).sum();
-  disp /= r1.rows();
-
-  for (int i = 0; i < r1.rows(); i++) {
-    r1(i, 0) += disp[0];
-    r1(i, 1) += disp[1];
-    r1(i, 2) += disp[2];
-  }
-
-  m1.setPositions(r1);
-  return;
-}
-
-void helper_functions::translationRemove(Matter &m1, const Matter &m2) {
-  AtomMatrix r2 = m2.getPositions();
-  translationRemove(m1, r2);
-  return;
-}
-
 double helper_functions::maxAtomMotion(const AtomMatrix v1) {
   double max = 0.0;
   for (int i = 0; i < v1.rows(); i++) {
@@ -557,19 +434,6 @@ AtomMatrix helper_functions::loadMode(std::string filename, int nAtoms) {
   return mode;
 }
 
-void helper_functions::saveMode(FILE *modeFile, std::shared_ptr<Matter> matter,
-                                AtomMatrix mode) {
-  long const nAtoms = matter->numberOfAtoms();
-  for (long i = 0; i < nAtoms; ++i) {
-    if (matter->getFixed(i)) {
-      fprintf(modeFile, "0 0 0\n");
-    } else {
-      fprintf(modeFile, "%lf\t%lf \t%lf\n", mode(i, 0), mode(i, 1), mode(i, 2));
-    }
-  }
-  return;
-}
-
 std::vector<int> helper_functions::split_string_int(std::string s,
                                                     std::string delim) {
   std::vector<int> list;
@@ -598,182 +462,6 @@ std::vector<int> helper_functions::split_string_int(std::string s,
   return list;
 }
 
-namespace helper_functions {
-struct atom {
-  double r;
-  int z;
-};
-
-struct by_atom {
-  bool operator()(atom const &a, atom const &b) const {
-    if (a.z != b.z) {
-      return a.z < b.z;
-    } else {
-      return a.r < b.r;
-    }
-  }
-};
-} // namespace helper_functions
-
 double roundUp(double x, double f) { return ceil(x / f); }
-
-bool helper_functions::identical(const Matter &m1, const Matter &m2,
-                                 const double distanceDifference) {
-
-  AtomMatrix r1 = m1.getPositions();
-  AtomMatrix r2 = m2.getPositions();
-
-  std::set<int> matched;
-  double tolerance = distanceDifference;
-
-  if (r1.rows() != r2.rows()) {
-    return false;
-  }
-  int N = r1.rows();
-
-  for (int i = 0; i <= N; i++) {
-    if (fabs((m1.pbc(r1.row(i) - r2.row(i))).norm()) < tolerance &&
-        m1.getAtomicNr(i) == m2.getAtomicNr(i)) {
-      matched.insert(i);
-    }
-  }
-
-  for (int j = 0; j < N; j++) {
-
-    if (matched.count(j) == 1)
-      continue;
-
-    for (int k = 0; k < N; k++) {
-      if (matched.count(j) == 1)
-        break;
-
-      if (fabs((m1.pbc(r1.row(j) - r2.row(k))).norm()) < tolerance &&
-          m1.getAtomicNr(j) == m2.getAtomicNr(k)) {
-        matched.insert(j);
-      }
-    }
-
-    // XXX: can we abort early if no match was found?
-  }
-
-  if (matched.size() == (unsigned)N) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-bool helper_functions::sortedR(const Matter &m1, const Matter &m2,
-                               const double distanceDifference) {
-  SPDLOG_INFO("In sortedR");
-  AtomMatrix r1 = m1.getPositions();
-  AtomMatrix r2 = m2.getPositions();
-  double tolerance = distanceDifference;
-  int matches = 0;
-  if (r1.rows() != r2.rows()) {
-    return false;
-  }
-
-  // Allocate memory for rdf1 and rdf2
-  std::vector<std::set<atom, by_atom>> rdf1(r1.rows());
-  std::vector<std::set<atom, by_atom>> rdf2(r2.rows());
-
-  for (int i2 = 0; i2 < r2.rows(); i2++) {
-    rdf2[i2].clear();
-    for (int j2 = 0; j2 < r2.rows(); j2++) {
-      if (j2 == i2)
-        continue;
-      atom a2;
-      a2.r = m2.distance(i2, j2);
-      a2.z = m2.getAtomicNr(j2);
-      rdf2[i2].insert(a2);
-      rdf2[j2].insert(a2);
-    }
-  }
-  for (int i1 = 0; i1 < r1.rows(); i1++) {
-    if (matches == i1 - 2) {
-      return false;
-    }
-    for (int j1 = 0; j1 < r1.rows(); j1++) {
-      if (j1 == i1)
-        continue;
-      atom a;
-      a.r = m1.distance(i1, j1);
-      a.z = m1.getAtomicNr(j1);
-      rdf1[i1].insert(a);
-      rdf1[j1].insert(a);
-    }
-    for (int x = 0; x < r2.rows(); x++) {
-      auto it2 = rdf2[x].begin();
-      auto it = rdf1[i1].begin();
-      int c = 0;
-      int counter = 0;
-      for (; c < r1.rows(); c++) {
-        if (it == rdf1[i1].end() || it2 == rdf2[x].end())
-          break;
-        atom k1 = *it;
-        atom k2 = *it2;
-        if (fabs(k1.r - k2.r) < tolerance && k1.z == k2.z) {
-          counter++;
-        } else {
-          SPDLOG_INFO("No match");
-          break;
-        }
-        ++it;
-        ++it2;
-      }
-      if (counter == r1.rows()) {
-        matches++;
-      } else {
-        SPDLOG_INFO("No match");
-      }
-    }
-  }
-
-  return matches >= r1.rows();
-}
-
-void helper_functions::pushApart(std::shared_ptr<Matter> m1,
-                                 double minDistance) {
-  if (minDistance <= 0)
-    return;
-
-  AtomMatrix r1 = m1->getPositions();
-  MatrixType Force(r1.rows(), 3);
-  double f = 0.025;
-  double cut = minDistance;
-  double pushAparts = 500;
-  for (int p = 0; p < r1.rows(); p++) {
-    for (int axis = 0; axis <= 2; axis++) {
-      Force(p, axis) = 0;
-    }
-  }
-  for (int count = 0; count < pushAparts; count++) {
-    int moved = 0;
-    for (int i = 0; i < r1.rows(); i++) {
-      for (int j = i + 1; j < r1.rows(); j++) {
-        double d = m1->distance(i, j);
-        if (d < cut) {
-          moved++;
-          for (int axis = 0; axis <= 2; axis++) {
-            double componant = f * (r1(i, axis) - r1(j, axis)) / d;
-            Force(i, axis) += componant;
-            Force(j, axis) -= componant;
-          }
-        }
-      }
-    }
-    if (moved == 0)
-      break;
-    for (int k = 0; k < r1.rows(); k++) {
-      for (int axis = 0; axis <= 2; axis++) {
-        r1(k, axis) += Force(k, axis);
-        Force(k, axis) = 0;
-      }
-    }
-    m1->setPositions(r1);
-    // m1->matter2con("movie.con", true);
-  }
-}
 
 } // namespace eonc
