+++ client/HelperFunctions.h
@@ -11,7 +11,6 @@
 */
 #pragma once
 #include "Eigen.h"
-#include "client/matter/Matter.h"
 #include <string>
 #include <vector>
 
@@ -65,12 +64,6 @@ AtomMatrix makeOrthogonal(
 void makeProjection(double *result, const double *v1, const double *v2,
                     long size); // result = projection of v1 on v2
 RotationMatrix rotationExtract(const AtomMatrix r1, const AtomMatrix r2);
-bool rotationMatch(const Matter &m1, const Matter &m2, const double max_diff);
-void rotationRemove(const AtomMatrix r1, std::shared_ptr<Matter> m2);
-void rotationRemove(const std::shared_ptr<Matter> m1,
-                    std::shared_ptr<Matter> m2);
-void translationRemove(Matter &m1, const AtomMatrix r1);
-void translationRemove(Matter &m1, const Matter &m2);
 double maxAtomMotion(const AtomMatrix v1);
 double maxAtomMotionV(const VectorType v1);
 long numAtomsMoved(const AtomMatrix v1, double cutoff);
@@ -86,14 +79,7 @@ getRelevantFile(std::string filename); // return filename containing _checkpoint
 VectorType loadMasses(std::string filename, int nAtoms);
 AtomMatrix loadMode(FILE *modeFile, int nAtoms);
 AtomMatrix loadMode(std::string filename, int nAtoms);
-void saveMode(FILE *modeFile, std::shared_ptr<Matter> matter, AtomMatrix mode);
 std::vector<int> split_string_int(std::string s, std::string delim);
-
-bool identical(const Matter &m1, const Matter &m2,
-               const double distanceDifference);
-bool sortedR(const Matter &m1, const Matter &m2,
-             const double distanceDifference);
-void pushApart(std::shared_ptr<Matter> m1, double minDistance);
 
 } // namespace helper_functions
